move: 달력 로직을 :feature:main으로 이동하고 calendar 모듈을 삭제
