--- feature/calendar/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="academic_calendar_top_bar_title">학사 일정</string>
-    <string name="calendar_btn_navigate_previous">이전 달로 이동</string>
-    <string name="calendar_btn_navigate_next">다음 달로 이동</string>
-</resources>
