+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/calendar/compose/component/AcademicCalendarTopBar.kt
@@ -1,12 +1,12 @@
-package com.ku_stacks.ku_ring.calendar.compose.component
+package com.ku_stacks.ku_ring.main.calendar.compose.component
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import com.ku_stacks.ku_ring.calendar.R.string.academic_calendar_top_bar_title
 import com.ku_stacks.ku_ring.designsystem.components.LightAndDarkPreview
 import com.ku_stacks.ku_ring.designsystem.components.topbar.CenterTitleTopBar
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.KuringTheme
+import com.ku_stacks.ku_ring.main.R.string.academic_calendar_top_bar_title
 
 @Composable
 internal fun AcademicCalendarTopBar(
