+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/calendar/compose/component/calendar/CalendarColor.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.calendar.compose.component.calendar
+package com.ku_stacks.ku_ring.main.calendar.compose.component.calendar
 
 import androidx.compose.ui.graphics.Color
