+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/calendar/type/ScheduleType.kt
@@ -1,11 +1,7 @@
-package com.ku_stacks.ku_ring.calendar.type
+package com.ku_stacks.ku_ring.main.calendar.type
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
-import com.ku_stacks.ku_ring.calendar.type.ScheduleType.DEGREE
-import com.ku_stacks.ku_ring.calendar.type.ScheduleType.ETC
-import com.ku_stacks.ku_ring.calendar.type.ScheduleType.EVENT
-import com.ku_stacks.ku_ring.calendar.type.ScheduleType.REGISTRATION
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.KuringTheme
 
 
@@ -24,8 +20,8 @@ enum class ScheduleType {
 
 @Composable
 fun ScheduleType.color(): Color = when(this) {
-    DEGREE -> KuringTheme.colors.degree
-    REGISTRATION -> KuringTheme.colors.registration
-    EVENT -> KuringTheme.colors.event
-    ETC -> KuringTheme.colors.etc
+    ScheduleType.DEGREE -> KuringTheme.colors.degree
+    ScheduleType.REGISTRATION -> KuringTheme.colors.registration
+    ScheduleType.EVENT -> KuringTheme.colors.event
+    ScheduleType.ETC -> KuringTheme.colors.etc
 }
